+++ src/perception/launch/debug.launch.py
@@ -1,0 +1,17 @@
+from launch import LaunchDescription
+from launch.substitutions import LaunchConfiguration
+from launch_ros.actions import Node
+from launch.actions import DeclareLaunchArgument
+
+
+def generate_launch_description():
+    return LaunchDescription(
+        [
+            Node(
+                package="perception",
+                executable="perception",
+                name="perception",
+                arguments=["--ros-args", "--log-level", "perception:=debug"],
+            ),
+        ]
+    )
